ToolBar Position Backup/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyMuPDF==1.23.5
easyocr==1.7.0
Pillow==10.0.1
Werkzeug==2.3.7
numpy==1.24.3
torch==2.0.1
opencv-python==4.8.1.78
bijoy2unicode==0.1.1
